requirements-dev.txt
# Core
pydantic>=2.9.2
pydantic-settings>=2.5.2
PyJWT==2.8.0
python-dotenv==1.0.0

# AWS
boto3==1.34.0
botocore==1.34.0
boto3-stubs[dynamodb,sns,lambda,sagemaker-runtime]
aws-cdk-lib
constructs
moto[dynamodb,s3,sns,sqs,lambda]==5.0.0

# HTTP
requests==2.31.0
httpx==0.26.0

# Redis cache
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.92.2

# ML
numpy>=2.1.3
# tensorflow (optional for training – heavy install)
# tensorflow==2.15.0

# Dev tools
black==23.12.1
isort==5.13.2
flake8==7.0.0
